Added other two-way walls
